3.9 实现 BVH 优化, 随机选择轴对基元排序(not work)

+++ src/aabb.rs
@@ -42,7 +42,7 @@ impl AABB {
         }
     }
 
-    pub fn axis_interval(&self, n: i32) -> &Interval {
+    pub fn axis_interval(&self, n: usize) -> &Interval {
         match n {
             1 => &self.y,
             2 => &self.z,
